requirements.txt
prompt-toolkit
websocket-client
pyotp
rsa
requests
